% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{area_scale_raster}
\alias{area_scale_raster}
\title{area-scale raster, max 1}
\usage{
area_scale_raster(r, r_a)
}
\arguments{
\item{r}{raster to scale}

\item{r_a}{area raster}
}
\value{
raster multiplied by area and divided by the maximum value to max at 1
}
\description{
area-scale raster, max 1
}
