% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bio_aquamaps}
\alias{bio_aquamaps}
\title{Global marine biodiversity indicators from AquaMaps}
\format{A \code{\link[raster]{stack}} with indicator (\code{nspp_*} or \code{rls_*}) and taxonomic group or all taxa (\code{*_all}).}
\source{
\url{https://github.com/marinebon/bbnj}
}
\usage{
bio_aquamaps
}
\description{
Dataset containing global half-degree rasters clipped to the high seas of
species richness (\code{nspp_*}) and Red List Sum (\code{rls_*})  package
\href{https://github.com/marinebon/bbnj}{marinebon/bbnj}.
}
\keyword{datasets}
