% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_d_prjres}
\alias{get_d_prjres}
\title{Get dataset object in given projection-resolution}
\usage{
get_d_prjres(dataset, prjres = "", debug = F)
}
\arguments{
\item{dataset}{see datasets}

\item{prjres}{projection-resolution, per projections_tbl$prjres}

\item{debug}{defaults to False}
}
\value{
polygon (s), raster (r), or stack (s) per prefix of dataset in given projection-resolution (prjres)
}
\description{
Get dataset object in given projection-resolution
}
