% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{get_grid}
\alias{get_grid}
\title{Get empty global grid}
\usage{
get_grid(res = 0.1, crs = leaflet:::epsg4326)
}
\arguments{
\item{res}{resolution, in decimal degrees (default 0.1)}

\item{crs}{coordinate resolutions system, ie projection (default geographic coordinate system 4326, ie WGS84)}
}
\value{
global raster with extent -180 to 180 longitude, -90 to 90 latitude and specified resolution
}
\description{
Get empty global grid
}
\examples{

# get half-degree global grid
get_grid(res=0.5)
}
