% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_r_projection}
\alias{get_r_projection}
\title{Get projection info for raster}
\usage{
get_r_projection(r, debug = F)
}
\arguments{
\item{r}{raster}
}
\value{
character string of projection_tbl$prjres
}
\description{
Get projection info for raster
}
