% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{get_tif_area_stats}
\alias{get_tif_area_stats}
\title{get area statistics from tif of solution}
\usage{
get_tif_area_stats(tif)
}
\arguments{
\item{tif}{}
}
\value{
named list of area stats for high seas (\code{!is.na(value)}) and solution (\code{value=1})
}
\description{
get area statistics from tif of solution
}
