% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_tif_projection}
\alias{get_tif_projection}
\title{Get projection info for tif of raster}
\usage{
get_tif_projection(tif)
}
\arguments{
\item{tif}{path to raster tif}
}
\value{
character string of projection_tbl$prjres
}
\description{
Get projection info for tif of raster
}
