% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{get_tif_solution}
\alias{get_tif_solution}
\title{get raster solution of tif}
\usage{
get_tif_solution(tif)
}
\arguments{
\item{tif}{}
}
\value{
raster solution of tif with attributes appended
}
\description{
get raster solution of tif
}
