% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{lyr_to_tif}
\alias{lyr_to_tif}
\title{Write layer in raster stack to tif}
\usage{
lyr_to_tif(lyr, s, s_dir, debug = F)
}
\arguments{
\item{lyr}{layer name}

\item{s}{stack}

\item{s_dir}{directory name}

\item{debug}{defaults to False}
}
\description{
Write layer in raster stack to tif
}
