% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_abnj}
\alias{p_abnj}
\title{Polygon of areas beyond national jurisdiction, ie high seas area}
\format{A \code{\link[sf]{sf}} object of a polygon.}
\usage{
p_abnj
}
\description{
Global area [-180,180,-90,90] with land and Exclusive Economic Zones (EEZs)
clipped out using the "Intersect_EEZ_IHO_v3_2018" product from \href{http://MarineRegions.org}{MarineRegions.org}.
}
\keyword{datasets}
