% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_abnj_ppow}
\alias{p_abnj_ppow}
\title{Polygon of Pelagic Provinces of the World (Spalding et al, 2012) for high
seas}
\format{A \code{\link[sf]{sf}} object of a polygon.}
\usage{
p_abnj_ppow
}
\description{
See http://data.unep-wcmc.org/datasets/38. Clipped to high seas. TODO: merge
small areas with larger ones.
}
\keyword{datasets}
