% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_eez}
\alias{p_eez}
\title{Polygons of exclusive economic zones (EEZs)}
\format{A \code{\link[sf]{sf}} object of a polygon.}
\usage{
p_eez
}
\description{
Downloaded World_EEZ_v10_20180221/eez_v10.shp from MarineRegions.org.
}
\keyword{datasets}
