% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_eez_s05}
\alias{p_eez_s05}
\title{Simplified polygons of exclusive economic zones (EEZs)}
\format{A \code{\link[sf]{sf}} object of a polygon.}
\usage{
p_eez_s05
}
\description{
Downloaded World_EEZ_v10_20180221/eez_v10.shp from \href{http://MarineRegions.org}{MarineRegions.org}. For
quick mapping applied simplification using
\code{\link[rmapshaper]{ms_simplify}(keep = 0.5)} twice.
}
\keyword{datasets}
