% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_highseas}
\alias{p_highseas}
\title{Polygon of high seas area}
\format{A \code{\link[sf]{sf}} object of a polygon.}
\usage{
p_highseas
}
\description{
Global area [-180,180,-90,90] with land and Exclusive Economic Zones (EEZs
frommMarineRegions.org) clipped out.
}
\keyword{datasets}
