% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_iho}
\alias{p_iho}
\title{Polygon of IHO Seas}
\format{
A \code{\link[sf]{sf}} object of a polygon.
}
\usage{
p_iho
}
\description{
See \url{http://www.marineregions.org/sources.php#iho}. Clipped to high seas. TODO:
merge small areas with larger ones.
}
\keyword{datasets}
