% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_iho_s05}
\alias{p_iho_s05}
\title{Simplified polygon of IHO Seas}
\format{
A \code{\link[sf]{sf}} object of a polygon.
}
\usage{
p_iho_s05
}
\description{
See \url{http://www.marineregions.org/sources.php#iho}. Clipped to high seas. TODO:
merge small areas with larger ones. Simplified to 5% of original vertices for fast visualization.
}
\keyword{datasets}
