% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_ihor_s05}
\alias{p_ihor_s05}
\title{Simplified polygon of IHO Seas Revised, ie Seven Seas}
\format{
A \code{\link[sf]{sf}} object of a polygon.
}
\usage{
p_ihor_s05
}
\description{
See \url{http://www.marineregions.org/sources.php#iho}. Clipped to high seas.
Merged smaller areas to the top 7 seas by taking the nearest sea from the
centroid. Simplified to 5% of original vertices for fast visualization.
}
\keyword{datasets}
