% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_mine_claims}
\alias{p_mine_claims}
\title{Polygons of mine claims (ISA)}
\format{A \code{\link[sf]{sf}} object of mining claim polygons.}
\usage{
p_mine_claims
}
\description{
Per ISA_claim_areas_update_20181202.
}
\keyword{datasets}
