% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_ppow_s05}
\alias{p_ppow_s05}
\title{Simplified polygon of Pelagic Provinces of the World (Spalding et al, 2012) for high
seas}
\format{A \code{\link[sf]{sf}} object of a polygon.}
\usage{
p_ppow_s05
}
\description{
See \url{http://data.unep-wcmc.org/datasets/38}. Clipped to high seas. TODO: merge
small areas with larger ones. Simplified to 5% of original vertices for fast visualization.
}
\keyword{datasets}
