% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{problem_diagnostics}
\alias{problem_diagnostics}
\title{Produce diagnostics table for setting relative targets of features}
\usage{
problem_diagnostics(pu, features, budget)
}
\arguments{
\item{pu}{planning unit raster}

\item{features}{features stack}

\item{budget}{in percent of area}
}
\value{
table with rel_all and rel_each per feature
}
\description{
Produce diagnostics table for setting relative targets of features
}
