% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{r_diff}
\alias{r_diff}
\title{Calculate difference of two raster solutions}
\usage{
r_diff(r1, r2)
}
\arguments{
\item{r1}{1st raster solution [0,1]}

\item{r2}{2nd raster solution [0,1]}
}
\value{
r2 - r1
}
\description{
Calculate difference of two raster solutions
}
