% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_ihor}
\alias{r_ihor}
\title{Raster for polygon of IHO Seas Revised, ie Seven Seas}
\format{A \code{\link[raster]{raster}} prence (1 or 0) of mine claim in cell.}
\usage{
r_ihor
}
\description{
Global half-degree raster of seven seas in the high seas by seaid (alphabetical):
\enumerate{
  \item Arctic Ocean
  \item Indian Ocean
  \item North Atlantic Ocean
  \item North Pacific Ocean
  \item South Atlantic Ocean
  \item South Pacific Ocean
  \item Southern Ocean
}
}
\keyword{datasets}
