% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_mine_claims}
\alias{r_mine_claims}
\title{Raster for presence of mining claim (Interridge Vent Database v3.4)}
\format{A \code{\link[raster]{raster}} prence (1 or 0) of mine claim in cell.}
\usage{
r_mine_claims
}
\description{
Global half-degree raster of presence in high seas, per ISA_claim_areas_update_20181202.
}
\keyword{datasets}
