% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_phys_scapes_hetero}
\alias{r_phys_scapes_hetero}
\title{Raster of heterogeneity of 1 thru 11 classes of benthic seascapes (Harris
& Whiteway, 2009)}
\format{
An object of class \code{RasterLayer} of dimension 360 x 720 x 1.
}
\usage{
r_phys_scapes_hetero
}
\description{
Focal variety, ie number of unique classes, within a 20-cell radius. See \code{\link{s_phys_scapes}}.
}
\keyword{datasets}
