% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_phys_seamounts}
\alias{r_phys_seamounts}
\title{Raster for count of seamounts (Kim & Wessel, 2011)}
\format{A \code{\link[raster]{raster}} count of seamount features within cell.}
\usage{
r_phys_seamounts
}
\description{
Global half-degree raster of high seas.
}
\keyword{datasets}
