% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_phys_vents}
\alias{r_phys_vents}
\title{Raster for count of hydrothermal vents (Interridge Vent Database v3.4)}
\format{A \code{\link[raster]{raster}} count of hydrothermal vents features within cell.}
\usage{
r_phys_vents
}
\description{
Global half-degree raster of high seas.
}
\keyword{datasets}
