% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_pu_id}
\alias{r_pu_id}
\title{Raster of planning unit id for high seas}
\format{A \code{\link[raster]{raster}} with integer of id unique to cell.}
\usage{
r_pu_id
}
\description{
Global half-degree raster of high seas with unique id.
}
\keyword{datasets}
