% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{r_to_prjres}
\alias{r_to_prjres}
\title{Project raster to projection-raster}
\usage{
r_to_prjres(prjres, r, name, method = c("bilinear", "ngb"), debug = F)
}
\arguments{
\item{prjres}{projection-resolution, per projections_tbl$prjres}

\item{r}{raster}

\item{name}{dataset name without type prefix (p_|r_|s_)}

\item{method}{defaults to bilinear, or nearest neighbor (ngb)}

\item{debug}{defaults to False}
}
\description{
Project raster to projection-raster
}
