% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_vgpm}
\alias{r_vgpm}
\title{Raster for average Vertically Generalized Production Model (VGPM)}
\format{A \code{\link[raster]{raster}} of half-degree cells}
\usage{
r_vgpm
}
\description{
Global half-degree raster of mean primary production using the Vertically
Generalized Production Model (Standard VGPM 2013-02-01 to 2019-01-31) using
\href{Standard VGPM - 2160 by 4320 Monthly XYZ files from VIIRS
Data}{http://orca.science.oregonstate.edu/2160.by.4320.monthly.xyz.vgpm.v.chl.v.sst.php}
}
\keyword{datasets}
