% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{report_solution}
\alias{report_solution}
\title{Report on conservation solution}
\usage{
report_solution(tif)
}
\arguments{
\item{tif}{path to raster tif of solution}
}
\value{
for use in reports to spit out area stats, map and table of resulting scenario
}
\description{
Report on conservation solution
}
