% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{rescale_raster}
\alias{rescale_raster}
\title{Rescale raster}
\usage{
rescale_raster(r, rescale = T, log = F, inverse = F,
  multiply_area = F)
}
\arguments{
\item{r}{input raster}

\item{rescale}{T/F to rescale values 0 to 1, default=T}

\item{log}{T/F to log transform, defaul=F}

\item{inverse}{T/F to invert by subtracting from 1}

\item{multiply_area}{T/F to multiply raster by 1}
}
\value{
raster
}
\description{
Rescale raster
}
