% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{rescale_stack}
\alias{rescale_stack}
\title{Rescale stack}
\usage{
rescale_stack(s, from_all = T, rescale = T, log = F, inverse = F,
  multiply_area = F)
}
\arguments{
\item{s}{stack}

\item{from_all}{rescale based on range from all rasters in stack
(default=TRUE) or within each raster (FALSE)}

\item{rescale}{T/F to rescale values 0 to 1, default=T}

\item{log}{T/F to log transform, defaul=F}

\item{inverse}{T/F to invert by subtracting from 1}

\item{multiply_area}{T/F to multiply raster by 1}
}
\value{
raster
}
\description{
Rescale stack
}
