% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny app map viewer for BBNJ inputs and}
\usage{
run_app()
}
\value{
Launches the app locally in your web browser. If \code{inst/app} directory
  exists in your current path (as when working from the R project of the bbnj
  repository) it will launch that version.
}
\description{
Run the Shiny app map viewer for BBNJ inputs and
}
