% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_bio_gmbi}
\alias{s_bio_gmbi}
\title{Raster stack of global marine biodiversity indicators from AquaMaps and IUCN}
\format{A \code{\link[raster]{stack}} with indicator (\code{nspp_*} or \code{rls_*}) and taxonomic group or all taxa (\code{*_all}).}
\source{
\url{https://github.com/marinebon/bbnj}

\url{https://aquamaps.org}

\url{https://iucnredlist.org}
}
\usage{
s_bio_gmbi
}
\description{
Dataset containing global half-degree rasters clipped to the high seas of
species richness (\code{nspp_*}) and Red List Sum (\code{rls_*})  package
\href{https://github.com/marinebon/bbnj}{marinebon/bbnj}.
}
\keyword{datasets}
