% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_fish_saup_v2}
\alias{s_fish_saup_v2}
\title{Raster stack version 2 from Sea Around Us Project at UBC (Cheung, Lam et al;
in draft) of maximum catch potential}
\format{A \code{\link[raster]{stack}} with layers of results from analyses.}
\usage{
s_fish_saup_v2
}
\description{
Global half-degree raster of high seas. Maximum catch potential (MCP;
landings in metric tons) of more than 1,000 fish and invertebrates species
for \code{mcp_2004} (average 1995 to 2014) and  \code{mcp_2050} (average 2041
to 2060 under 'business as usual' climate change scenario GFDL 8.5).
}
\details{
Version 2: This pattern along Southeast Pacific in the original version is due to the exceptional
high catch potential of Chilean jack mackerel in the upwelling region.
Indeed, most of our maps have to adjust by removing this species. Otherwise,
the high MCP in this region will mask the MCP of all other areas. I have re-extracted the data for you by
removing the MCP of this species. -- Vicky Lam
}
\keyword{datasets}
