% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_fish_ubc}
\alias{s_fish_ubc}
\title{Raster stack from UBC (Cheung, Lam et al; in draft) of maximum catch
potential}
\format{A \code{\link[raster]{stack}} with layers of results from analyses.}
\usage{
s_fish_ubc
}
\description{
Global half-degree raster of high seas. Maximum catch potential (MCP;
landings in metric tons) of more than 1,000 fish and invertebrates species
for \code{mcp_2004} (average 1995 to 2014) and  \code{mcp_2050} (average 2041 to
2060 under 'business as usual' climate change scenario GFDL 8.5).
}
\keyword{datasets}
