% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_ihor}
\alias{s_ihor}
\title{Stack of rasters describing presence of seven seas from IHO revised (IHOR)}
\format{
A \code{\link[raster]{stack}} of \code{seaid} [1-7] with raster
  attribute table providing \code{sea} and \code{area_km2} (see
  \code{\link[raster]{factorValues}()}).
}
\usage{
s_ihor
}
\description{
Global half-degree stack of seven seas in the high seas by seaid (alphabetical):
\enumerate{
  \item Arctic Ocean
  \item Indian Ocean
  \item North Atlantic Ocean
  \item North Pacific Ocean
  \item South Atlantic Ocean
  \item South Pacific Ocean
  \item Southern Ocean
}
}
\keyword{datasets}
