% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_phys_scapes}
\alias{s_phys_scapes}
\title{Raster stack for area (km2) of 1 thru 11 classes of benthic seascapes (Harris
& Whiteway, 2009)}
\format{A \code{\link[raster]{stack}} in which each layer represents area for given integer class id of benthic seascape corresponding to :
 \enumerate{
  \item Abyssal, hilly plains, large (arched) uplifted structures, flat abyssal plains, flat, very low PP, very thin sediment (seascape 11)
  \item Abyssal, plains with slightly undulating seafloor, flat abyssal plains, continental rise, very flat, high DO, low PP, very cold (seascape 10)
  \item Abyssal, flat sedimented plains of marginal seas, central rift zone, ridge flanks, marginal seas with hilly bottoms, flat, low DO, thin sediment (seascape 8)
  \item Abyssal, volcanic ridges and highs, central rift zone, ridge flanks, microcontinents, cold (seascape 7)
  \item Abyssal (Hadal), trenches controlled by fracture zones, deep water trenches, large arched uplifted structures, low PP thin sediment, cold (seascape 9)
  \item Lower Bathyal (Abyssal-Hadal), deep water trenches, island arcs, trenches controlled by fracture zones, volcanic ridges and plateaus, very steep (seascape 6)
  \item Lower Bathyal, continental slope see high PP, very thick sediment, warm (seascape 4)
  \item Lower Bathyal, other ridges and plateaus, marginal plateaus, marginal seas with hilly bottom, steep, very low DO (seascape 3)
  \item Lower Bathyal, deep shelf (submerged), marginal plateaus, very high DO, high PP, thick sediment, warm (seascape 2)
  \item Upper Bathyal, shallow shelf, low DO, very high PP. thick sediment very warm (seascape 1)
  \item Lower Bathyal, island arcs, steep, high DO (seascape 5)
}

Note that seascape numbers differ from those in Harris & Whiteway 2009. Numbers in parenthesis correspond to numbers from Harris & Whiteway 2009.}
\usage{
s_phys_scapes
}
\description{
Global half-degree raster of high seas. Original 0.1 deg cells were
multiplied by area and summed to 0.5 deg cells.
}
\keyword{datasets}
