% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{s_to_prjres}
\alias{s_to_prjres}
\title{Project stack to projection-raster}
\usage{
s_to_prjres(prjres, s, name, method = c("bilinear", "ngb"), debug = F)
}
\arguments{
\item{prjres}{projection-resolution, per projections_tbl$prjres}

\item{s}{stack object}

\item{name}{dataset name without type prefix (p_|r_|s_)}

\item{method}{defaults to bilinear, or nearest neighbor (ngb)}

\item{debug}{defaults to False}
}
\description{
Project stack to projection-raster
}
