% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{scenarios_diff_png}
\alias{scenarios_diff_png}
\title{calculate scenario difference}
\usage{
scenarios_diff_png(scenarios, dir_scenarios, dir_diffs)
}
\arguments{
\item{scenarios}{a named list with scenarios}

\item{dir_scenarios}{directory to find scenarios}

\item{dir_diffs}{directory to output png difference}
}
\description{
calculate scenario difference
}
