% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{solve_log}
\alias{solve_log}
\title{Helper function to run prioritizr::solve() and log output and resulting raster}
\usage{
solve_log(p, pfx = deparse(substitute(p)))
}
\arguments{
\item{p}{\code{\link[prioritizr]{problem}} to feed into \code{\link[prioritizr]{solve}}}

\item{log}{path to output log from solving problem}
}
\value{
output from \code{\link[prioritizr]{solve}}
}
\description{
Helper function to run prioritizr::solve() and log output and resulting raster
}
