% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritizr.R
\name{tbl_target_representation}
\alias{tbl_target_representation}
\title{Display table of target representaiton in conservation problem solution}
\usage{
tbl_target_representation(csv = glue("{pfx}_rep.csv"))
}
\arguments{
\item{csv}{comma-seperated value output from \code{\link{solve_log}}}
}
\value{
\code{\link[formattable]{formattable}} in form of
  \code{\link[DT]{datatable}} with horizontal bars indicating percent held
  per target.
}
\description{
Display table of target representaiton in conservation problem solution
}
