% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{tif_to_shp_gcs}
\alias{tif_to_shp_gcs}
\title{Convert raster tif to polygon shapefile in geographic coordinates for mapping
prioritizr solution in non-Mercator projection}
\usage{
tif_to_shp_gcs(tif)
}
\arguments{
\item{tif}{path to raster tif}
}
\value{
True if successfully writes shapefile to same path as tif (.tif ->
  _gcs.shp)
}
\description{
Convert raster tif to polygon shapefile in geographic coordinates for mapping
prioritizr solution in non-Mercator projection
}
