% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_phys_scapes}
\alias{s_phys_scapes}
\title{Raster stack for area (km2) of 1 thru 11 classes of benthic seascapes (Harris
& Whiteway, 2009)}
\format{A \code{\link[raster]{raster}} integer class id of benthic seascape.}
\usage{
s_phys_scapes
}
\description{
Global half-degree raster of high seas. Original 0.1 deg cells were
multiplied by area and summed to 0.5 deg cells.
}
\keyword{datasets}
